Created converter program for sprite group header files.
The header files were in the incorrect format for memcpy. The format works well for 8x8 sprites, but it seems to be incorrect for 16x16 sprites - the order in which values are read results in a mangled sprite. The additional program found in  will take in a header as input and naively convert to the correct format as long as the input is valid. It will likely break with different inputs. Oh well.